chore: rename utils to handlerUtil
