+++ src/CollectionsHandler.js
@@ -1,4 +1,4 @@
-import { handler } from './utils';
+import { handler } from './handlerUtil';
 
 const RDF = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#';
